backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
python-dotenv
groq
numpy
pydantic
openai
langchain-text-splitters
sqlalchemy
psycopg2-binary
alembic
